+++ tests/integration/test_norm.py
@@ -2,39 +2,57 @@
 Test correctness of matrix norm for shell matrices.
 '''
 
-import unittest as ut
 import numpy as np
+from scipy.sparse.linalg import norm
 import hamiltonians
 
 import dynamite_test_runner as dtr
 
 from dynamite import config
-from dynamite._backend.bbuild import have_gpu_shell
 from dynamite.tools import complex_enabled
+from dynamite.subspaces import SpinConserve, Parity
+
 
 class Hamiltonians(dtr.DynamiteTestCase):
 
-    def test_shell(self):
-
-        config._initialize()
-        from petsc4py import PETSc
-
+    def test_norm(self):
         for H_name in hamiltonians.get_names(complex_enabled()):
             if H_name == 'syk' and self.skip_flags['small_only']:
                 continue
-            with self.subTest(H = H_name):
-                H = getattr(hamiltonians, H_name)()
-                H.shell = True
-                petsc_norm = H.get_mat().norm(PETSc.NormType.INFINITY)
+            for subspace_name in ('full', 'spinconserve', 'mix'):
+                with self.subTest(H=H_name, subspace=subspace_name):
+                    H = getattr(hamiltonians, H_name)()
 
-                H.shell = False
-                shell_norm = H.get_mat().norm(PETSc.NormType.INFINITY)
+                    if subspace_name == 'spinconserve':
+                        H.subspace = SpinConserve(config.L, config.L//3)
+                        H.allow_projection = True
+                    elif subspace_name == 'mix':
+                        H.add_subspace(
+                            SpinConserve(config.L, config.L//3),
+                            Parity('even')
+                        )
+                        H.allow_projection = True
 
-                # the extra factor of 1E2 error is because PETSc ignores rounding
-                # errors when they compute their norm!!
-                eps = H.nnz * np.finfo(np.complex128).eps * 1E2
-                self.assertLess(np.abs(petsc_norm-shell_norm), eps,
-                                msg = '\npetsc: %e\nshell: %e' % (petsc_norm, shell_norm))
+                    H.shell = True
+                    petsc_norm = H.infinity_norm()
+
+                    H.shell = False
+                    shell_norm = H.infinity_norm()
+
+                    # the extra factor of 1E2 error is because PETSc ignores
+                    # floating point errors when they compute their norm!!
+                    eps = H.nnz * np.finfo(np.complex128).eps * 1E2
+
+                    msg = f'\npetsc: {petsc_norm}\nshell: {shell_norm}'
+                    self.assertLess(np.abs(petsc_norm-shell_norm), eps,
+                                    msg=msg)
+
+                    if not self.skip_flags['medium_only']:
+                        numpy_norm = norm(H.to_numpy(), ord=np.inf)
+                        msg = f'\nnumpy: {numpy_norm}\nshell: {shell_norm}'
+                        self.assertLess(np.abs(numpy_norm-shell_norm), eps,
+                                        msg=msg)
+
 
 if __name__ == '__main__':
     dtr.main()