+++ CHANGELOG.md
@@ -11,6 +11,7 @@
  - More descriptive output when `str()` is called on `states.State` class, and LaTeX representation of states in Jupyter
  - `State.set_all_by_function()` which takes a user-supplied function and applies it to set each element of a state vector
  - `State.set_uniform()` to set the state to a uniform superposition
+ - `Operator.infinity_norm()`
 
 ### Changed
  - Moved `operators.load_from_file` into the Operator class, as `operators.Operator.load`